Revert "Hot module reload fixes (#181)"
This reverts commit b59298ec0e3f4a70c1b5aa4e128778d803793572.
